// scripts/test-rewards/src/contracts/HydroBase.types.ts
export type Timestamp = Uint64;
export type Uint64 = string;
export type Binary = string;
export type Uint128 = string;
export type Decimal = string;
export type TokenInfoProviderInstantiateMsg = {
  lsm: {
    hub_connection_id: string;
    hub_transfer_channel_id: string;
    icq_update_period: number;
    max_validator_shares_participating: number;
  };
} | {
  token_info_provider_contract: {
    admin?: string | null;
    code_id: number;
    label: string;
    msg: Binary;
  };
};
// ...
export interface CollectionInfo {
  name: string;
  symbol: string;
}
export interface InstantiateContractMsg {
  admin?: string | null;
  code_id: number;
  label: string;
  msg: Binary;
}
export interface TrancheInfo {
  metadata: string;
  name: string;
}
export type ExecuteMsg = {
  lock_tokens: {
    lock_duration: number;
    proof?: LockTokensProof | null;
  };
} | {
  refresh_lock_duration: {
    lock_duration: number;
    lock_ids: number[];
  };
} | {
  unlock_tokens: {
    lock_ids?: number[] | null;
  };
} | {
  create_proposal: {
    deployment_duration: number;
    description: string;
    minimum_atom_liquidity_request: Uint128;
    round_id?: number | null;
    title: string;
    tranche_id: number;
  };
} | {
  vote: {
    proposals_votes: ProposalToLockups[];
    tranche_id: number;
  };
} | {
  unvote: {
    lock_ids: number[];
    tranche_id: number;
  };
} | {
  add_account_to_whitelist: {
    address: string;
  };
} | {
  remove_account_from_whitelist: {
    address: string;
  };
} | {
  update_config: {
    activate_at: Timestamp;
    cw721_collection_info?: CollectionInfo | null;
    known_users_cap?: number | null;
    max_deployment_duration?: number | null;
    max_locked_tokens?: number | null;
  };
} | {
  delete_configs: {
    timestamps: Timestamp[];
  };
} | {
  pause: {};
} | {
  add_tranche: {
    tranche: TrancheInfo;
  };
} | {
  edit_tranche: {
    tranche_id: number;
    tranche_metadata?: string | null;
    tranche_name?: string | null;
  };
} | {
  create_icqs_for_validators: {
    validators: string[];
  };
} | {
  add_i_c_q_manager: {
    address: string;
  };
} | {
  remove_i_c_q_manager: {
    address: string;
  };
} | {
  withdraw_i_c_q_funds: {
    amount: Uint128;
  };
} | {
  add_liquidity_deployment: {
    deployed_funds: Coin[];
    destinations: string[];
    funds_before_deployment: Coin[];
    proposal_id: number;
    remaining_rounds: number;
    round_id: number;
    total_rounds: number;
    tranche_id: number;
  };
} | {
  remove_liquidity_deployment: {
    proposal_id: number;
    round_id: number;
    tranche_id: number;
  };
} | {
  update_token_group_ratio: {
    new_ratio: Decimal;
    old_ratio: Decimal;
    token_group_id: string;
  };
} | {
  add_token_info_provider: {
    token_info_provider: TokenInfoProviderInstantiateMsg;
  };
} | {
  remove_token_info_provider: {
    provider_id: string;
  };
} | {
  set_gatekeeper: {
    gatekeeper_addr?: string | null;
  };
} | {
  transfer_nft: {
    recipient: string;
    token_id: string;
  };
} | {
  send_nft: {
    contract: string;
    msg: Binary;
    token_id: string;
  };
} | {
  lock_tokens_then_send_nft: {
    contract: string;
    lock_duration: number;
    msg: Binary;
    proof?: LockTokensProof | null;
  };
} | {
  approve: {
    expires?: Expiration | null;
    spender: string;
    token_id: string;
  };
} | {
  revoke: {
    spender: string;
    token_id: string;
  };
} | {
  approve_all: {
    expires?: Expiration | null;
    operator: string;
  };
} | {
  revoke_all: {
    operator: string;
  };
};
export type Expiration = {
  at_height: number;
} | {
  at_time: Timestamp;
} | {
  never: {};
};
export interface LockTokensProof {
  maximum_amount: Uint128;
  proof: string[];
  sig_info?: SignatureInfo | null;
}
export interface SignatureInfo {
  claim_msg: Binary;
  signature: Binary;
}
export interface ProposalToLockups {
  lock_ids: number[];
  proposal_id: number;
}
export interface Coin {
  amount: Uint128;
  denom: string;
}